# packages/api-contracts/rest/openapi.yaml
openapi: 3.1.0
info:
  title: ManaBandhu REST API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /posts:
    get:
      operationId: listCommunityPosts
      responses:
        '200':
          description: Community posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommunityPost'
    post:
      operationId: createCommunityPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostInput'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPost'
  /immigration/resources:
    get:
      operationId: listImmigrationResources
      responses:
        '200':
          description: Immigration resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImmigrationResource'
      parameters:
        - name: category
          in: query
          schema: { type: string }
    post:
      operationId: createImmigrationResource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImmigrationResourceInput'
      responses:
        '201':
          description: Resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImmigrationResource'
  /immigration/guides:
    get:
      operationId: listImmigrationGuides
      responses:
        '200':
          description: Immigration guides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImmigrationGuide'
      parameters:
        - name: category
          in: query
          schema: { type: string }
  /immigration/checklists:
    get:
      operationId: listImmigrationChecklists
      responses:
        '200':
          description: Immigration checklists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImmigrationChecklist'
      parameters:
        - name: category
          in: query
          schema: { type: string }
  /immigration/faq:
    get:
      operationId: listFaqItems
      responses:
        '200':
          description: FAQ items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FaqItem'
      parameters:
        - name: category
          in: query
          schema: { type: string }
...
  /rooms:
    get:
      operationId: listRoomListings
      parameters:
        - name: location
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Room listings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RoomListing' }
    post:
      operationId: createRoomListing
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRoomListingInput' }
      responses:
        '201':
          description: Listing created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomListing' }
...
  /rides:
    get:
      operationId: listRideOffers
      parameters:
        - name: origin
          in: query
          schema: { type: string }
        - name: destination
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Ride offers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RideOffer' }
    post:
      operationId: createRideOffer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRideOfferInput' }
      responses:
        '201':
          description: Ride offer created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RideOffer' }
...
  /rides/{rideId}/rate:
    post:
      operationId: rateRide
      parameters:
        - name: rideId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRideRatingInput' }
      responses:
        '201':
          description: Rating created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RideRating' }
...
  /posts/communities:
    get:
      operationId: listCommunities
      responses:
        '200':
          description: Communities
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Community' }
    post:
      operationId: createCommunity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCommunityInput' }
      responses:
        '201':
          description: Community created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Community' }
  /posts/{postId}/comments:
    get:
      operationId: listPostComments
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Post comments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PostComment' }
    post:
      operationId: createPostComment
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePostCommentInput' }
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostComment' }
  /posts/{postId}/reactions:
    get:
      operationId: listPostReactions
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Post reactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PostReaction' }
    post:
      operationId: createPostReaction
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePostReactionInput' }
      responses:
        '201':
          description: Reaction created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostReaction' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreatePostInput:
      type: object
      additionalProperties: false
      required: [title, body]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        body: { type: string, minLength: 1, maxLength: 4000 }
    CommunityPost:
      allOf:
        - $ref: '#/components/schemas/CreatePostInput'
        - type: object
          required: [id, ownerId, createdAt]
          properties:
            id: { type: string, format: uuid }
            ownerId: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
    ImmigrationResource:
      type: object
      additionalProperties: false
      required: [id, ownerId, title, description, category, resourceType, verified, createdAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        category: { type: string, maxLength: 80 }
        url: { type: string, format: uri, maxLength: 500 }
        resourceType: { type: string, enum: [DOCUMENT, GUIDE, LINK, FORM, TEMPLATE] }
        tags: { type: string, maxLength: 400 }
        verified: { type: boolean }
        createdAt: { type: string, format: date-time }
    CreateImmigrationResourceInput:
      type: object
      additionalProperties: false
      required: [title, description, category, resourceType]
      properties:
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        category: { type: string, maxLength: 80 }
        url: { type: string, format: uri, maxLength: 500 }
        resourceType: { type: string, enum: [DOCUMENT, GUIDE, LINK, FORM, TEMPLATE] }
        tags: { type: string, maxLength: 400 }
        verified: { type: boolean }
    ImmigrationGuide:
      type: object
      additionalProperties: false
      required: [id, ownerId, title, description, content, category, difficultyLevel, estimatedDurationMinutes, createdAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        content: { type: string, maxLength: 4000 }
        category: { type: string, maxLength: 80 }
        difficultyLevel: { type: string, enum: [BEGINNER, INTERMEDIATE, ADVANCED] }
        estimatedDurationMinutes: { type: integer }
        createdAt: { type: string, format: date-time }
    ImmigrationChecklist:
      type: object
      additionalProperties: false
      required: [id, ownerId, title, description, category, itemsJson, createdAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        category: { type: string, maxLength: 80 }
        itemsJson: { type: string, maxLength: 4000 }
        dueDate: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...
    FaqItem:
      type: object
      additionalProperties: false
      required: [id, question, answer, category, published, sortOrder, createdAt]
      properties:
        id: { type: string, format: uuid }
        question: { type: string, maxLength: 500 }
        answer: { type: string, maxLength: 4000 }
        category: { type: string, maxLength: 80 }
        tags: { type: string, maxLength: 400 }
        published: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
...
    RoomListing:
      type: object
      additionalProperties: false
      required: [id, ownerId, title, price, roomType, status, broadLocation, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        price: { type: number, format: double }
        roomType: { type: string, enum: [PRIVATE, SHARED, ENTIRE] }
        status: { type: string, enum: [DRAFT, ACTIVE, PAUSED, RENTED] }
        broadLocation: { type: string, maxLength: 200 }
        exactAddress: { type: string, maxLength: 4000 }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateRoomListingInput:
      type: object
      additionalProperties: false
      required: [title, price, roomType, broadLocation]
      properties:
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        price: { type: number, format: double }
        roomType: { type: string, enum: [PRIVATE, SHARED, ENTIRE] }
        broadLocation: { type: string, maxLength: 200 }
        exactAddress: { type: string, maxLength: 4000 }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
...
    RideOffer:
      type: object
      additionalProperties: false
      required: [id, driverId, originArea, destinationArea, departureAt, seatsTotal, seatsAvailable, contribution, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        driverId: { type: string, format: uuid }
        originArea: { type: string, maxLength: 200 }
        destinationArea: { type: string, maxLength: 200 }
        departureAt: { type: string, format: date-time }
        seatsTotal: { type: integer }
        seatsAvailable: { type: integer }
        contribution: { type: string, maxLength: 200 }
        status: { type: string, enum: [OPEN, FULL, COMPLETED, CANCELED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateRideOfferInput:
      type: object
      additionalProperties: false
      required: [originArea, destinationArea, departureAt, seatsTotal, contribution]
      properties:
        originArea: { type: string, maxLength: 200 }
        destinationArea: { type: string, maxLength: 200 }
        departureAt: { type: string, format: date-time }
        seatsTotal: { type: integer }
        contribution: { type: string, maxLength: 200 }
...
    CreateRideRatingInput:
      type: object
      additionalProperties: false
      required: [rateeId, rating]
      properties:
        rateeId: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, maxLength: 4000 }
    RideRating:
      type: object
      additionalProperties: false
      required: [id, rideId, raterId, rateeId, rating, createdAt]
      properties:
        id: { type: string, format: uuid }
        rideId: { type: string, format: uuid }
        raterId: { type: string, format: uuid }
        rateeId: { type: string, format: uuid }
        rating: { type: integer }
        comment: { type: string, maxLength: 4000 }
        createdAt: { type: string, format: date-time }
    Community:
      type: object
      additionalProperties: false
      required: [id, ownerId, name, createdAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        name: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
        createdAt: { type: string, format: date-time }
    CreateCommunityInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 4000 }
    PostComment:
      type: object
      additionalProperties: false
      required: [id, postId, authorId, body, createdAt]
      properties:
        id: { type: string, format: uuid }
        postId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        body: { type: string, maxLength: 4000 }
        createdAt: { type: string, format: date-time }
    CreatePostCommentInput:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body: { type: string, maxLength: 4000 }
    PostReaction:
      type: object
      additionalProperties: false
      required: [id, postId, userId, type, createdAt]
      properties:
        id: { type: string, format: uuid }
        postId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        type: { type: string, maxLength: 20 }
        createdAt: { type: string, format: date-time }
    CreatePostReactionInput:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type: { type: string, maxLength: 20 }
...
security:
  - bearerAuth: []